[Radoub] tech-debt: Split ModelPreviewGLControl.cs into partials (#2127)
Extract the cleanly-separable concerns from the 1987-line control into partials:
.Input (pointer/keyboard + screen-to-world camera gestures), .Animation
(playback + pose sampling), .Emitters (particle-system lifecycle + emitter
world transforms), .Textures (PLT color/texture management + GPU upload). The
GL init/deinit/render loop and mesh-buffer build (tightly coupled, shared
locals) stay in the main file alongside shared state and SetPreviewState. The
pre-existing .Particles partial is unchanged. Pure mechanical split, no behavior change.
Tests: Radoub.UI.Tests 1360/1360; full solution builds 0 errors.
