+++ Radoub.UI/Radoub.UI/Controls/ModelPreviewGLControl.cs
@@ -283,114 +283,6 @@ public partial class ModelPreviewGLControl : OpenGlControlBase
     }
 
     /// <summary>
-    /// Set the texture service for loading textures.
-    /// </summary>
-    public void SetTextureService(TextureService textureService)
-    {
-        _textureService = textureService;
-        _needsTextureUpdate = true;
-        RequestNextFrameRendering();
-    }
-
-    /// <summary>
-    /// Set character colors for PLT texture rendering.
-    /// </summary>
-    public void SetCharacterColors(int skinColor, int hairColor, int tattoo1Color, int tattoo2Color)
-    {
-        SetColorIndices(new PltColorIndices
-        {
-            Skin = skinColor,
-            Hair = hairColor,
-            Tattoo1 = tattoo1Color,
-            Tattoo2 = tattoo2Color,
-            Metal1 = _colorIndices.Metal1,
-            Metal2 = _colorIndices.Metal2,
-            Cloth1 = _colorIndices.Cloth1,
-            Cloth2 = _colorIndices.Cloth2,
-            Leather1 = _colorIndices.Leather1,
-            Leather2 = _colorIndices.Leather2
-        });
-    }
-
-    /// <summary>
-    /// Set all PLT color indices (body + armor colors).
-    /// </summary>
-    public void SetColorIndices(PltColorIndices colorIndices)
-    {
-        if (!ColorsEqual(_colorIndices, colorIndices))
-        {
-            _colorIndices = colorIndices;
-            // Only PLT textures depend on color indices. Flat TGA/DDS textures
-            // are color-independent and can survive the change (#2058).
-            InvalidatePltTextures();
-            _needsTextureUpdate = true;
-            RequestNextFrameRendering();
-        }
-    }
-
-    private void InvalidatePltTextures()
-    {
-        if (_pltTextureNames.Count == 0) return;
-
-        foreach (var name in _pltTextureNames)
-        {
-            if (_textureCache.TryGetValue(name, out var texId))
-            {
-                if (_gl != null && texId != 0)
-                    _gl.DeleteTexture(texId);
-                _textureCache.Remove(name);
-                _textureAlphaProfile.Remove(name);
-            }
-        }
-        _pltTextureNames.Clear();
-    }
-
-    /// <summary>
-    /// Set armor colors for PLT texture rendering.
-    /// </summary>
-    public void SetArmorColors(int metal1, int metal2, int cloth1, int cloth2, int leather1, int leather2)
-    {
-        SetColorIndices(new PltColorIndices
-        {
-            Skin = _colorIndices.Skin,
-            Hair = _colorIndices.Hair,
-            Tattoo1 = _colorIndices.Tattoo1,
-            Tattoo2 = _colorIndices.Tattoo2,
-            Metal1 = metal1,
-            Metal2 = metal2,
-            Cloth1 = cloth1,
-            Cloth2 = cloth2,
-            Leather1 = leather1,
-            Leather2 = leather2
-        });
-    }
-
-    /// <summary>
-    /// Clears the GPU texture cache so textures reload with current color indices.
-    /// Also called on module switch to clear stale HAK textures (#1867).
-    /// </summary>
-    public void ClearTextureCache()
-    {
-        if (_gl != null)
-        {
-            foreach (var texId in _textureCache.Values)
-                _gl.DeleteTexture(texId);
-        }
-        _textureCache.Clear();
-        _textureAlphaProfile.Clear();
-        _pltTextureNames.Clear();
-    }
-
-    private static bool ColorsEqual(PltColorIndices a, PltColorIndices b)
-    {
-        return a.Skin == b.Skin && a.Hair == b.Hair &&
-               a.Tattoo1 == b.Tattoo1 && a.Tattoo2 == b.Tattoo2 &&
-               a.Metal1 == b.Metal1 && a.Metal2 == b.Metal2 &&
-               a.Cloth1 == b.Cloth1 && a.Cloth2 == b.Cloth2 &&
-               a.Leather1 == b.Leather1 && a.Leather2 == b.Leather2;
-    }
-
-    /// <summary>
     /// Rotate the model incrementally.
     /// </summary>
     public void Rotate(float deltaY, float deltaX = 0)
@@ -406,575 +298,6 @@ public partial class ModelPreviewGLControl : OpenGlControlBase
     {
         _viewController.ResetView();
         RequestNextFrameRendering();
-    }
-
-    // ----- Pointer / keyboard input (#2124) -----
-    //
-    // OpenGlControlBase has no Background brush, so pointer events don't
-    // hit-test on its transparent pixels. The host panel overlays a
-    // transparent Border on top of this control and forwards input here
-    // via the public Handle* / Pan / ZoomAtCursor / RotateBy methods
-    // below. The OnPointerPressed etc. overrides remain as a fallback
-    // if the control is ever used without the overlay.
-
-    /// <summary>
-    /// Host panels call this to request a pan in screen-pixel units.
-    /// Converts to world units using the last-rendered viewport + camera distance.
-    /// </summary>
-    public void PanByPixels(double dxPixels, double dyPixels)
-    {
-        if (_lastViewportHeight <= 0) return;
-        float fovRadians = MathF.PI / 6f;
-        float halfFovTan = MathF.Tan(fovRadians * 0.5f);
-        float worldPerPixel = 2f * _lastCameraDistance * halfFovTan / _lastViewportHeight;
-
-        // See OnPointerMoved for axis convention; camera right = -X, up = +Z.
-        _viewController.Pan(new Vector3(-(float)dxPixels * worldPerPixel, 0, (float)dyPixels * worldPerPixel));
-        RequestNextFrameRendering();
-    }
-
-    /// <summary>
-    /// Host panels call this to zoom toward a world point under the cursor.
-    /// </summary>
-    public void ZoomAtCursorPixels(Point screenPos, double wheelDeltaY)
-    {
-        float factor = (float)Math.Pow(1.1, wheelDeltaY);
-        var pivot = UnprojectToTargetPlane(screenPos);
-        _viewController.ZoomAtPoint(factor, pivot);
-        RequestNextFrameRendering();
-    }
-
-    /// <summary>
-    /// Host panels call this to rotate by screen-pixel drag deltas.
-    /// </summary>
-    public void RotateByPixels(double dxPixels, double dyPixels)
-    {
-        _viewController.Rotate((float)(dxPixels * 0.01), (float)(dyPixels * 0.01));
-        RequestNextFrameRendering();
-    }
-
-    /// <summary>
-    /// Cycle debug visualisation mode (0..4).
-    /// </summary>
-    public void CycleDebugMode()
-    {
-        DebugMode = (_debugMode + 1) % 5;
-    }
-
-    /// <summary>
-    /// Snap to a preset view (front/side/top/back). #2124.
-    /// </summary>
-    public void SetViewPreset(ViewPreset preset)
-    {
-        _viewController.SetViewPreset(preset);
-        RequestNextFrameRendering();
-    }
-
-    // ----- Animation playback (#2124) -----
-
-    /// <summary>
-    /// Currently active animation, or null if none selected.
-    /// </summary>
-    public MdlAnimation? ActiveAnimation => _activeAnimation;
-
-    /// <summary>
-    /// Current playhead time in seconds (0..Animation.Length).
-    /// </summary>
-    public float AnimationTime
-    {
-        get => _animTime;
-        set
-        {
-            _animTime = value;
-            _cachedPose = null;
-            _needsMeshUpdate = true;
-            RequestNextFrameRendering();
-        }
-    }
-
-    /// <summary>
-    /// True when animation is auto-advancing. Toggle with Play/Pause.
-    /// </summary>
-    public bool IsAnimationPlaying => _animPlaying;
-
-    /// <summary>
-    /// Playback speed multiplier (1.0 = real-time).
-    /// </summary>
-    public float AnimationSpeed
-    {
-        get => _animSpeed;
-        set => _animSpeed = Math.Clamp(value, 0.1f, 5f);
-    }
-
-    /// <summary>
-    /// Select an animation by reference (null to clear). Resets playhead to 0.
-    /// </summary>
-    public void SetActiveAnimation(MdlAnimation? animation)
-    {
-        _activeAnimation = animation;
-        _animTime = 0f;
-        _cachedPose = null;
-        _needsMeshUpdate = true;
-        RequestNextFrameRendering();
-    }
-
-    public void PlayAnimation()
-    {
-        if (_activeAnimation == null) return;
-        _animPlaying = true;
-        _animLastTick = DateTime.UtcNow;
-        EnsureAnimTimer();
-    }
-
-    public void PauseAnimation()
-    {
-        _animPlaying = false;
-    }
-
-    private void EnsureAnimTimer()
-    {
-        if (_animTimer != null) return;
-        _animTimer = new Avalonia.Threading.DispatcherTimer
-        {
-            Interval = TimeSpan.FromMilliseconds(33), // ~30 fps
-        };
-        _animTimer.Tick += OnAnimTick;
-        _animTimer.Start();
-    }
-
-    private void OnAnimTick(object? sender, EventArgs e)
-    {
-        // Particle simulation runs BEFORE the skeletal-animation early-return (#2395)
-        // so emitters animate even when no animation is selected/playing (the common
-        // case for a static blueprint preview). Never let a sim exception escape into
-        // the timer/render loop (CLAUDE.md) — log WARN and continue.
-        if (_particleSystems.Count > 0)
-        {
-            try
-            {
-                UpdateParticleSystems();
-            }
-            catch (Exception ex)
-            {
-                UnifiedLogger.LogApplication(LogLevel.WARN,
-                    $"[Particle] tick update failed (non-fatal): {ex.GetType().Name}: {ex.Message}");
-            }
-        }
-
-        if (!_animPlaying || _activeAnimation == null) return;
-
-        var now = DateTime.UtcNow;
-        var dt = (float)(now - _animLastTick).TotalSeconds * _animSpeed;
-        _animLastTick = now;
-
-        float length = _activeAnimation.Length;
-        if (length <= 0)
-        {
-            _animTime = 0f;
-        }
-        else
-        {
-            _animTime += dt;
-            if (_animTime >= length)
-                _animTime -= length * MathF.Floor(_animTime / length); // loop
-        }
-
-        _cachedPose = null;
-        _needsMeshUpdate = true;
-        RequestNextFrameRendering();
-    }
-
-    /// <summary>
-    /// (Re)build the per-emitter <see cref="Radoub.UI.Particles.ParticleSystem"/> list for a
-    /// newly-assigned model (#2395). Clears any existing systems, then compiles one system per
-    /// <see cref="MdlEmitterNode"/>. Each system gets a stable, distinct seed (index+1) so
-    /// identical emitters don't move in lockstep. If any emitter exists, resets the particle
-    /// clock and ensures the animation timer runs so the sim advances even with no skeletal
-    /// animation selected.
-    /// </summary>
-    private void RebuildParticleSystems(MdlModel? model)
-    {
-        _particleSystems.Clear();
-
-        if (model == null || !model.HasEmitterNodes())
-            return;
-
-        int index = 0;
-        foreach (var node in model.EnumerateAllNodes())
-        {
-            if (node is not MdlEmitterNode emitter) continue;
-            try
-            {
-                var compiled = Radoub.UI.Particles.EmitterCompiler.Compile(emitter);
-
-                // At-rest gate (the #2439 root): placeable destruction effects (wood debris, fire)
-                // are emitters keyed to a "die" animation — their birthrate is 0 in the default
-                // pose and they only fire when destroyed. The preview shows the placeable at rest,
-                // so those emitters must be silent. The model leads: the signal is the emitter's
-                // birthrate in the "default" animation, not the static header peak. Continuous
-                // emitters (brazier flame, creature dust) keep a non-zero default birthrate and
-                // render. Full animation-driven emission (firing during a played "die") is a
-                // follow-up. (#2544 / #2439)
-                if (!Radoub.UI.Particles.EmitterAnimationGate.ShouldRenderAtRest(model, emitter.Name))
-                {
-                    UnifiedLogger.LogApplication(LogLevel.INFO,
-                        $"[Particle] Emitter '{emitter.Name}' is gated to a non-default animation (silent at rest) — not rendered. (#2544)");
-                    continue;
-                }
-
-                // Secondary, model-led cull: an emitter authored with no effective output (birthrate
-                // ≤ 0 or lifespan ≤ 0) produces nothing — skip building a system for it. (#2544)
-                if (emitter.BirthRate <= 0f || emitter.LifeExp <= 0f)
-                {
-                    UnifiedLogger.LogApplication(LogLevel.INFO,
-                        $"[Particle] Emitter '{emitter.Name}' has no effective output (birthRate={emitter.BirthRate}, lifeExp={emitter.LifeExp}) — not rendered. (#2544)");
-                    continue;
-                }
-
-                // Render modes with no faithful implementation (Beam/Mesh/LinkedChain) still fall
-                // back to a camera-facing billboard — log once per model so the limitation stays visible.
-                bool unsupportedRender = compiled.RenderMode == Radoub.UI.Particles.ParticleRenderMode.Beam
-                    || compiled.RenderMode == Radoub.UI.Particles.ParticleRenderMode.Mesh
-                    || compiled.RenderMode == Radoub.UI.Particles.ParticleRenderMode.LinkedChain;
-                if (unsupportedRender)
-                {
-                    UnifiedLogger.LogApplication(LogLevel.INFO,
-                        $"[Particle] Emitter '{emitter.Name}' render mode {compiled.RenderMode} not faithfully supported — rendered as a camera-facing billboard (approximate). (#2544)");
-                }
-
-                var system = new Radoub.UI.Particles.ParticleSystem(compiled, (uint)(index + 1));
-                // Pre-warm to steady-state so the emitter looks already-running on first frame
-                // (Aurora pre-warms; otherwise particles fill in slowly from empty). (#2395)
-                system.PreWarm(GetEmitterWorldPosition(emitter), GetEmitterWorldRotation(emitter));
-                _particleSystems.Add((emitter, compiled, system));
-                index++;
-            }
-            catch (Exception ex)
-            {
-                UnifiedLogger.LogApplication(LogLevel.WARN,
-                    $"[Particle] failed to compile emitter '{emitter.Name}' (skipped): {ex.GetType().Name}: {ex.Message}");
-            }
-        }
-
-        if (_particleSystems.Count > 0)
-        {
-            _particleLastTick = DateTime.UtcNow;
-            EnsureAnimTimer();
-            UnifiedLogger.LogApplication(LogLevel.INFO,
-                $"[Particle] built {_particleSystems.Count} particle system(s) for model '{model.Name}'");
-        }
-    }
-
-    /// <summary>
-    /// True if the active model has any emitter node that is animated by some animation (i.e. an
-    /// ancestor or the emitter itself has position/orientation keyframes). Tools use this to decide
-    /// whether to default the preview to a playing idle animation so animated emitters (e.g. the
-    /// fairy's flapping wing Dummies) sweep correctly instead of rendering as a static orb. (#2434)
-    /// </summary>
-    public bool HasAnimatedEmitters()
-    {
-        if (_model == null || _model.Animations.Count == 0) return false;
-        if (!_model.HasEmitterNodes()) return false;
-
-        // Collect the geometry-tree ancestor names of every emitter.
-        var emitterChainNames = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        foreach (var node in _model.EnumerateAllNodes())
-        {
-            if (node is not MdlEmitterNode) continue;
-            for (var cur = (MdlNode?)node; cur != null; cur = cur.Parent)
-                if (!string.IsNullOrEmpty(cur.Name)) emitterChainNames.Add(cur.Name);
-        }
-
-        // An emitter is animated if any animation keyframes a node in its chain.
-        foreach (var anim in _model.Animations)
-        {
-            if (anim.GeometryRoot == null) continue;
-            if (AnimTreeKeyframesAny(anim.GeometryRoot, emitterChainNames)) return true;
-        }
-        return false;
-    }
-
-    private static bool AnimTreeKeyframesAny(MdlNode animNode, HashSet<string> names)
-    {
-        bool keyed = animNode.PositionTimes.Length > 1 || animNode.OrientationTimes.Length > 1
-            || animNode.ScaleTimes.Length > 1;
-        if (keyed && !string.IsNullOrEmpty(animNode.Name) && names.Contains(animNode.Name))
-            return true;
-        foreach (var child in animNode.Children)
-            if (AnimTreeKeyframesAny(child, names)) return true;
-        return false;
-    }
-
-    /// <summary>
-    /// Advance every particle system by the wall-clock delta since the last particle tick (#2395).
-    /// ParticleSystem already guards/clamps dt internally, so no extra clamp here. Throttled
-    /// DEBUG log (~1s) reports system count + total live particles for the UAT log-smoke check.
-    /// </summary>
-    private void UpdateParticleSystems()
-    {
-        var now = DateTime.UtcNow;
-        float pdt = (float)(now - _particleLastTick).TotalSeconds;
-        _particleLastTick = now;
-
-        // Sample the active animation's pose so emitter nodes follow animated parents (e.g. the
-        // fairy's wing Dummies flap during idle — without this the wing emitter sits at its static
-        // bind pose and its particles stack into an orb instead of sweeping into a wing blade). Null
-        // pose (no animation playing) falls back to the static bind transform. (#2434)
-        var pose = GetCurrentPose();
-        foreach (var (node, _, system) in _particleSystems)
-        {
-            var worldPos = GetEmitterWorldPosition(node, pose);
-            var worldRot = GetEmitterWorldRotation(node, pose);
-            system.Update(pdt, worldPos, worldRot);
-        }
-
-        RequestNextFrameRendering();
-
-        // Throttled diagnostic (~1s) so UAT can confirm particles are alive without log spam.
-        if ((now - _lastParticleLogTime).TotalSeconds >= 1.0)
-        {
-            _lastParticleLogTime = now;
-            int liveTotal = 0;
-            foreach (var (_, _, system) in _particleSystems)
-                liveTotal += system.LiveCount;
-            UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                $"[Particle] systems={_particleSystems.Count} liveTotal={liveTotal}");
-        }
-    }
-
-    /// <summary>
-    /// World-space spawn origin for an emitter node (#2395). Reuses the same node-hierarchy
-    /// world-transform plumbing the mesh path uses (<see cref="ModelViewController.GetWorldTransform(MdlNode?)"/>,
-    /// which composes Position/Orientation/Scale up the Parent chain), transforming the local
-    /// origin to world space. NOTE: the mesh buffer additionally pre-centers vertices by the
-    /// model's geometric center (UpdateMeshBuffersCore Pass 2); particle positions are NOT yet
-    /// centered to match, so the sim runs in raw model space. This is fine for sim-only wiring —
-    /// the next (render) task will reconcile particle coords with the centered mesh frame.
-    /// </summary>
-    private static Vector3 GetEmitterWorldPosition(MdlEmitterNode node,
-        IReadOnlyDictionary<string, ModelViewController.NodePose>? pose = null)
-    {
-        var world = ModelViewController.GetWorldTransform(node, pose);
-        return ModelViewController.TransformPosition(Vector3.Zero, world);
-    }
-
-    /// <summary>
-    /// World-space rotation for an emitter node (#2395). Mirrors <see cref="GetEmitterWorldPosition"/>:
-    /// composes the node's world transform up the Parent chain, then extracts the rotation as a
-    /// Quaternion. Decomposes (not raw <c>CreateFromRotationMatrix</c>) so any scale in the
-    /// transform doesn't corrupt the rotation; falls back to identity if decomposition fails.
-    /// </summary>
-    private static Quaternion GetEmitterWorldRotation(MdlEmitterNode node,
-        IReadOnlyDictionary<string, ModelViewController.NodePose>? pose = null)
-    {
-        var world = ModelViewController.GetWorldTransform(node, pose);
-        return Matrix4x4.Decompose(world, out _, out var rotation, out _)
-            ? rotation
-            : Quaternion.Identity;
-    }
-
-    /// <summary>
-    /// Build a node-name → sampled-pose dictionary from the active animation
-    /// at the current <see cref="AnimationTime"/>. Cached until playhead moves
-    /// or the animation changes.
-    /// </summary>
-    private Dictionary<string, ModelViewController.NodePose>? GetCurrentPose()
-    {
-        if (_activeAnimation?.GeometryRoot == null) return null;
-        if (_cachedPose != null) return _cachedPose;
-
-        var pose = new Dictionary<string, ModelViewController.NodePose>(StringComparer.OrdinalIgnoreCase);
-        BuildPoseRecursive(_activeAnimation.GeometryRoot, _animTime, pose);
-        _cachedPose = pose;
-        return pose;
-    }
-
-    private static void BuildPoseRecursive(MdlNode animNode, float t,
-        Dictionary<string, ModelViewController.NodePose> pose)
-    {
-        bool hasPos = animNode.PositionTimes.Length > 1;
-        bool hasOri = animNode.OrientationTimes.Length > 1;
-        bool hasScl = animNode.ScaleTimes.Length > 1;
-
-        if (hasPos || hasOri || hasScl)
-        {
-            var p = new ModelViewController.NodePose(
-                hasPos, hasPos ? MdlAnimationEvaluator.EvaluatePosition(animNode, t) : animNode.Position,
-                hasOri, hasOri ? MdlAnimationEvaluator.EvaluateOrientation(animNode, t) : animNode.Orientation,
-                hasScl, hasScl ? MdlAnimationEvaluator.EvaluateScale(animNode, t) : animNode.Scale);
-            if (!string.IsNullOrEmpty(animNode.Name))
-                pose[animNode.Name] = p;
-        }
-
-        foreach (var child in animNode.Children)
-            BuildPoseRecursive(child, t, pose);
-    }
-
-    protected override void OnPointerPressed(PointerPressedEventArgs e)
-    {
-        base.OnPointerPressed(e);
-        var props = e.GetCurrentPoint(this).Properties;
-        bool shift = (e.KeyModifiers & KeyModifiers.Shift) != 0;
-
-        if (props.IsMiddleButtonPressed || (props.IsLeftButtonPressed && shift))
-        {
-            _dragMode = DragMode.Pan;
-        }
-        else if (props.IsLeftButtonPressed)
-        {
-            _dragMode = DragMode.Rotate;
-        }
-        else
-        {
-            return;
-        }
-
-        _lastPointerPos = e.GetPosition(this);
-        Focus();
-        e.Pointer.Capture(this);
-        e.Handled = true;
-    }
-
-    protected override void OnPointerMoved(PointerEventArgs e)
-    {
-        base.OnPointerMoved(e);
-        if (_dragMode == DragMode.None) return;
-
-        var pos = e.GetPosition(this);
-        double dx = pos.X - _lastPointerPos.X;
-        double dy = pos.Y - _lastPointerPos.Y;
-        _lastPointerPos = pos;
-
-        if (_dragMode == DragMode.Rotate)
-        {
-            // 0.01 rad/px matches the feel of the discrete rotate buttons
-            // (0.3 rad per click). Free X-axis orbit — no clamping.
-            _viewController.Rotate((float)(dx * 0.01), (float)(dy * 0.01));
-            RequestNextFrameRendering();
-        }
-        else if (_dragMode == DragMode.Pan)
-        {
-            // Convert screen-pixel delta to world units at the target's depth.
-            // At the camera-to-target distance, one NDC unit covers
-            // cameraDistance * halfFovTan world units vertically.
-            if (_lastViewportHeight <= 0) return;
-            float fovRadians = MathF.PI / 6f;
-            float halfFovTan = MathF.Tan(fovRadians * 0.5f);
-            float worldPerPixel = 2f * _lastCameraDistance * halfFovTan / _lastViewportHeight;
-
-            var worldDelta = ScreenDeltaToWorld(-(float)dx * worldPerPixel, (float)dy * worldPerPixel);
-            _viewController.Pan(worldDelta);
-            RequestNextFrameRendering();
-        }
-    }
-
-    protected override void OnPointerReleased(PointerReleasedEventArgs e)
-    {
-        base.OnPointerReleased(e);
-        if (_dragMode != DragMode.None)
-        {
-            _dragMode = DragMode.None;
-            e.Pointer.Capture(null);
-            e.Handled = true;
-        }
-    }
-
-    protected override void OnPointerWheelChanged(PointerWheelEventArgs e)
-    {
-        base.OnPointerWheelChanged(e);
-
-        // Each wheel notch is ~1.0 in e.Delta.Y. 1.1× per notch gives
-        // a natural zoom feel without being twitchy.
-        float factor = (float)Math.Pow(1.1, e.Delta.Y);
-
-        var pos = e.GetPosition(this);
-        var pivot = UnprojectToTargetPlane(pos);
-        _viewController.ZoomAtPoint(factor, pivot);
-
-        RequestNextFrameRendering();
-        e.Handled = true;
-    }
-
-    protected override void OnKeyDown(KeyEventArgs e)
-    {
-        base.OnKeyDown(e);
-        const float rotStep = 0.1f;
-
-        switch (e.Key)
-        {
-            case Key.Left:
-            case Key.A:
-                _viewController.Rotate(-rotStep, 0);
-                break;
-            case Key.Right:
-            case Key.D:
-                _viewController.Rotate(rotStep, 0);
-                break;
-            case Key.Up:
-            case Key.W:
-                _viewController.Rotate(0, -rotStep);
-                break;
-            case Key.Down:
-            case Key.S:
-                _viewController.Rotate(0, rotStep);
-                break;
-            case Key.Home:
-                _viewController.ResetView();
-                break;
-            case Key.F8:
-                DebugMode = (_debugMode + 1) % 5;
-                return; // DebugMode setter already triggers re-render
-            default:
-                return;
-        }
-
-        RequestNextFrameRendering();
-        e.Handled = true;
-    }
-
-    /// <summary>
-    /// Convert a screen-space (pixel) delta into a world-space delta in the
-    /// plane facing the camera, using the current rotation state. Pan drags
-    /// move the model along these in-plane axes rather than world XY so the
-    /// motion feels right at any view angle.
-    /// </summary>
-    private Vector3 ScreenDeltaToWorld(float rightAmount, float upAmount)
-    {
-        // The model rotation is applied to vertices, so the visible
-        // "right" / "up" directions in world space are the unrotated
-        // camera axes. Camera looks along +Y (LookAt from -Y eye to
-        // origin with Z-up), so "right" = -X and "up" = +Z when no
-        // rotation has been applied. We want pan to move the model
-        // under the mouse regardless of the model rotation — so we
-        // translate the camera target along world-aligned axes only.
-        // This matches common DCC tool behaviour.
-        return new Vector3(rightAmount, 0, upAmount);
-    }
-
-    /// <summary>
-    /// Approximate the world-space point under the cursor at the depth of
-    /// the camera target, so scroll-wheel zoom can pivot there.
-    /// </summary>
-    private Vector3 UnprojectToTargetPlane(Point screenPos)
-    {
-        if (_lastViewportWidth <= 0 || _lastViewportHeight <= 0)
-            return _viewController.CameraTarget;
-
-        // NDC coordinates (-1..1), Y inverted because screen Y is top-down.
-        float ndcX = (float)(screenPos.X / _lastViewportWidth * 2.0 - 1.0);
-        float ndcY = (float)(1.0 - screenPos.Y / _lastViewportHeight * 2.0);
-
-        // At the target's depth, one NDC unit vertically = cameraDistance * halfFovTan world units.
-        float fovRadians = MathF.PI / 6f;
-        float halfFovTan = MathF.Tan(fovRadians * 0.5f);
-        float aspect = (float)_lastViewportWidth / _lastViewportHeight;
-        float worldY = ndcY * _lastCameraDistance * halfFovTan;
-        float worldX = ndcX * _lastCameraDistance * halfFovTan * aspect;
-
-        // The camera's right/up axes at the target depth. See
-        // ScreenDeltaToWorld — right = -X, up = +Z under our LookAt.
-        var target = _viewController.CameraTarget;
-        return target + new Vector3(-worldX, 0, worldY);
     }
 
     protected override void OnOpenGlInit(GlInterface gl)
@@ -1753,218 +1076,6 @@ public partial class ModelPreviewGLControl : OpenGlControlBase
         }
 
         _gl.BindVertexArray(0);
-    }
-
-    private void UpdateTextures()
-    {
-        if (_gl == null || _textureService == null || _model == null) return;
-
-        _textureRemapping.Clear();
-
-        // Collect unique texture names from meshes, remembering the mesh's MTR material
-        // name (#2497) so the loader can resolve diffuse from the .mtr texture0 when the
-        // bare bitmap misses (the white-model case).
-        var textureNames = new HashSet<string>();
-        var materialByTexture = new Dictionary<string, string>();
-        foreach (var mesh in _model.GetMeshNodes())
-        {
-            // Only consider meshes that are actually drawn. Render=false bone/internal
-            // meshes carry bitmaps (often the model name, e.g. boar's hidden legs use
-            // 'c_boar') that resolve to a base-game stub no visible surface uses —
-            // loading them wasted work and falsely tripped the base-texture warning. (#2029)
-            if (!mesh.Render || mesh.Vertices.Length == 0 || mesh.Faces.Length == 0)
-                continue;
-            if (string.IsNullOrEmpty(mesh.Bitmap))
-                continue;
-            var bitmap = mesh.Bitmap.ToLowerInvariant();
-            textureNames.Add(bitmap);
-            if (!string.IsNullOrEmpty(mesh.MaterialName) && !materialByTexture.ContainsKey(bitmap))
-                materialByTexture[bitmap] = mesh.MaterialName.ToLowerInvariant();
-        }
-
-        // #2395: include emitter particle textures so they load into _textureCache
-        // (the particle render path looks them up by Material.Texture).
-        foreach (var (_, emitter, _) in _particleSystems)
-        {
-            var t = emitter.Material.Texture;
-            if (!string.IsNullOrEmpty(t) && !t.Equals("null", StringComparison.OrdinalIgnoreCase))
-                textureNames.Add(t.ToLowerInvariant());
-        }
-
-        // Delete old textures that are no longer needed
-        var toRemove = new List<string>();
-        foreach (var kvp in _textureCache)
-        {
-            if (!textureNames.Contains(kvp.Key))
-            {
-                _gl.DeleteTexture(kvp.Value);
-                toRemove.Add(kvp.Key);
-            }
-        }
-        foreach (var key in toRemove)
-        {
-            _textureCache.Remove(key);
-            _textureAlphaProfile.Remove(key);
-            _pltTextureNames.Remove(key);
-        }
-
-        // Load new textures
-        UnifiedLogger.LogApplication(LogLevel.DEBUG,
-            $"UpdateTextures: {textureNames.Count} textures, colors: skin={_colorIndices.Skin}, hair={_colorIndices.Hair}, " +
-            $"metal1={_colorIndices.Metal1}, metal2={_colorIndices.Metal2}, cloth1={_colorIndices.Cloth1}, cloth2={_colorIndices.Cloth2}");
-
-        var failedTextures = new List<string>();
-        foreach (var texName in textureNames)
-        {
-            if (_textureCache.ContainsKey(texName))
-                continue; // Already loaded
-
-            try
-            {
-                materialByTexture.TryGetValue(texName, out var materialName);
-                var textureData = _preferBifTextures
-                    ? _textureService.LoadTexturePreferBIFWithKind(texName, _colorIndices)
-                    : _textureService.LoadTextureWithKind(texName, materialName, _colorIndices);
-                if (textureData == null)
-                {
-                    UnifiedLogger.LogApplication(LogLevel.DEBUG, $"  Texture '{texName}' returned null");
-                    failedTextures.Add(texName);
-                    continue;
-                }
-
-                var (width, height, pixels, isPlt) = textureData.Value;
-                var texId = UploadTexture(width, height, pixels);
-                if (texId != 0)
-                {
-                    _textureCache[texName] = texId;
-                    // PLT skin (player/part-based heads, bodies) is opaque by definition (#2540): a
-                    // PLT's "alpha" byte is a palette-LAYER index (skin/hair/tattoo selector), NOT
-                    // transparency. The engine never treats it as transparent — and scanning the
-                    // rendered RGBA would mis-read those palette-derived alphas as a cutout/blend,
-                    // wrongly carving solid humanoid heads (the gnome-head regression). Force Opaque.
-                    _textureAlphaProfile[texName] = isPlt
-                        ? AlphaProfile.Opaque
-                        : MeshTransparency.AnalyzeAlphaProfile(pixels, width, height);
-                    if (isPlt) _pltTextureNames.Add(texName);
-                    UnifiedLogger.LogApplication(LogLevel.DEBUG, $"  Loaded texture '{texName}' ({width}x{height}) -> texId={texId}, isPlt={isPlt}");
-                }
-            }
-            catch (Exception ex)
-            {
-                UnifiedLogger.LogApplication(LogLevel.WARN, $"Failed to load texture '{texName}': {ex.Message}");
-                failedTextures.Add(texName);
-            }
-        }
-
-        // For textures that failed to load, try model name as fallback.
-        // NWN creature models often have meshes with bitmap set to node name (e.g. "torso_g")
-        // instead of an actual texture. The real texture is typically the model name (e.g. "c_curst2").
-        if (failedTextures.Count > 0)
-        {
-            var modelTexture = _model.Name?.ToLowerInvariant();
-            if (!string.IsNullOrEmpty(modelTexture) && !_textureCache.ContainsKey(modelTexture))
-            {
-                var fallbackData = _preferBifTextures
-                    ? _textureService.LoadTexturePreferBIFWithKind(modelTexture, _colorIndices)
-                    : _textureService.LoadTextureWithKind(modelTexture, _colorIndices);
-                if (fallbackData != null)
-                {
-                    var (w, h, px, isPlt) = fallbackData.Value;
-                    var fallbackId = UploadTexture(w, h, px);
-                    if (fallbackId != 0)
-                    {
-                        _textureCache[modelTexture] = fallbackId;
-                        // PLT skin alpha is a palette-layer index, never transparency (#2540) — force
-                        // Opaque so a PLT body/head is never carved or blended (see primary site above).
-                        _textureAlphaProfile[modelTexture] = isPlt
-                            ? AlphaProfile.Opaque
-                            : MeshTransparency.AnalyzeAlphaProfile(px, w, h);
-                        if (isPlt) _pltTextureNames.Add(modelTexture);
-                        UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                            $"  Loaded model fallback texture '{modelTexture}' ({w}x{h}) -> texId={fallbackId}, isPlt={isPlt}");
-                    }
-                }
-            }
-
-            // Map failed textures to model texture if it loaded successfully
-            if (!string.IsNullOrEmpty(modelTexture) && _textureCache.ContainsKey(modelTexture))
-            {
-                foreach (var failed in failedTextures)
-                {
-                    _textureRemapping[failed] = modelTexture;
-                    UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                        $"  Remapped texture '{failed}' -> '{modelTexture}' (model fallback)");
-                }
-            }
-        }
-
-        // #1758: warn only when a non-base model's actually-rendered texture resolves from
-        // base-game BIF instead of its own pack. Computed here over the final rendered
-        // texture set — NOT inside the load loop — so the result is deterministic regardless
-        // of which textures were already cached from a previous model. The old in-loop check
-        // skipped cached textures, making the warning flicker on/off across reloads. (#2029)
-        bool usedBaseFallback = !_preferBifTextures
-            && textureNames.Any(t => _textureService.ResolvesFromBase(t));
-
-        if (Dispatcher.UIThread.CheckAccess())
-            TextureSourceChanged?.Invoke(this, usedBaseFallback);
-        else
-            Dispatcher.UIThread.Post(() => TextureSourceChanged?.Invoke(this, usedBaseFallback));
-    }
-
-    private uint UploadTexture(int width, int height, byte[] rgba)
-    {
-        if (_gl == null) return 0;
-
-        var texId = _gl.GenTexture();
-        _gl.BindTexture(TextureTarget.Texture2D, texId);
-
-        // Upload RGBA data
-        _gl.TexImage2D(TextureTarget.Texture2D, 0, InternalFormat.Rgba8,
-            (uint)width, (uint)height, 0, PixelFormat.Rgba, PixelType.UnsignedByte,
-            new ReadOnlySpan<byte>(rgba));
-
-        // Set texture parameters
-        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureMinFilter, (int)TextureMinFilter.LinearMipmapLinear);
-        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureMagFilter, (int)TextureMagFilter.Linear);
-        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureWrapS, (int)TextureWrapMode.Repeat);
-        _gl.TexParameter(TextureTarget.Texture2D, TextureParameterName.TextureWrapT, (int)TextureWrapMode.Repeat);
-
-        // Generate mipmaps for better quality at distance
-        _gl.GenerateMipmap(TextureTarget.Texture2D);
-
-        _gl.BindTexture(TextureTarget.Texture2D, 0);
-
-        return texId;
-    }
-
-    /// <summary>
-    /// Determines if a texture name represents a body/skin texture that should render behind armor.
-    /// NWN body part textures follow naming conventions like:
-    /// - pXX_bodyYYY (player body parts)
-    /// - pXX_headYYY (player heads)
-    /// - cXX_bodyYYY (creature body parts)
-    /// Armor textures don't contain these patterns.
-    /// </summary>
-    private static bool IsBodyTexture(string? textureName)
-    {
-        if (string.IsNullOrEmpty(textureName))
-            return false;
-
-        var name = textureName.ToLowerInvariant();
-
-        // Body part patterns for player/creature models
-        // These render BEHIND armor to prevent z-fighting
-        return name.Contains("_body") ||
-               name.Contains("_head") ||
-               name.Contains("_neck") ||
-               name.Contains("_hand") ||
-               name.Contains("_foot") ||
-               name.Contains("_shin") ||
-               name.Contains("_thigh") ||
-               name.Contains("_bicep") ||
-               name.Contains("_forearm") ||
-               name.Contains("_pelvis");
     }
 
     /// <summary>
